Pass mongodb URL directly to mongo
This lets users pass arbitrary query parameters without us getting in
the way, and also enables support for the nonstandard multiple host
syntax.

It's a little unclear on when the command line client learned to accept
URL arguments, so to maximize backwards-compatibility, use the old
multi-argument version when we can.