Accept buffer number passed to db#resolve()
This is another step in the direction of making the result buffer
synonymous with the result itself.

The dance with s:buffers is more than a bit awkward, but it enables
things to work when the buffer is unloaded.  I expect a more natural
solution will present itself when async support is rolled out.