Postgresql for windows
In my work PC on windows I have no rights for add 'psql' to path and should write full path to it.
I decided that it will be good idea to save psql engine to variable. But at the moment I don't know how to modify it in .vimrc file

How do you think, is it good idea?