Fix -E parameter conflict with AAD authentication
When no username is passed to the sqlserver connection URL, the adapter
adds the -E parameter to the sqlcmd command. However, when using
ActiveDirectoryDefault in the --authentication-method parameter, it
doesn't work because the go-sqlcmd program will choose integrated
authentication instead of AAD authentication.

Integrated authentication is chosen in go-sqlcmd either because the -E
parameter is provided or because the username is empty and the
authentication method is not specified.

This commit prevents the -E parameter from being passed if the
authentication method parameter is provided, ensuring that AAD
authentication can be used properly.